Very preliminary detector simulation adapted from T2K ND280.
This is an initial adaptation of the T2K ND280 detector geometry and
particle propagation simulation.  It has been slightly reorganized to
reflect "modern" G4 practices, and had the names changed (to protect the
innocent).  The current geometry is wrong.  I've added in very rough
approximations, and it's just good enough to test the event handling
infrastructure.

As of now, the geometry includes the wire planes (X,U,V), and the liquid
argon.  The initial version does not include any sensitive regions, which
will be included as part of the work on the electronics simulation and data
format.

Take *NOTHING* in this version as final.