Bug fix: Fix detsim-control check for /run/beamOn in the macro file.
The previous check was just looking for the presence of beamOn in the
input macro file, but it could have been in a comment.  The new check now
looks for the full macro command, and makes sure that it is anchored to the
start of the line (G4 macros must start in the firs column).