Match the wire geometry to what is in miniCAPTAIN
There has been a little bit of confusion about the actual miniCAPTAIN tpc
wire geometry, but it seems to be sorted out now that the detector is
installed and collecting data (unless there is a later commit fixing it
again).  The TPC has 1005 wires instrumented, and they are apparently
centered on the apex of the hexagon.