Add DETSIM specific short-circuit to NEST and only sum number of quanta.
DETSIM needs to number of electrons and photons generated per segment, but
does not need for the opticalphoton and thermalelectrons to be generated.
The NEST code has been modified to do this (inside a short ifdef) and the
information is being used to fill TG4HitSegment.  The hit segment is
backward compatible, so it can work with older version of elecSim.