Add the new DSimDokeBirks "recombination process".
NEST does a very good detailed calculation of the ionization and
scintillation yield in argon, but is more detailed than we need and quite
slow.  The DSimDokeBirks process copies the model that is most important
for a LArTPC, and removes a lot of the generality.  In particular, it
doesn't implement the Thomas-Imel model that becomes important at extremely
low energies.  As a result the DSimDokeBirks process is 200+ times faster
than NEST.  The new model should be used with care and results should be checked against the more accurate NEST model.