Be a little smarter about which trajectory points are saved.
This is a small change in how the trajectory points are selected for
output.  Previously, only particles which directly deposited energy had
trajectory points, now any particle that has a child which deposited energy
is saved.  Also, trajectory points where there is a hard interaction are
always selected.  Finally, for particles that directly deposit energy
(e.g. muons, but not neutrons), enough points are saved so that the
interpolated trajectory points stay within GetTrajectoryPointAccuracy().