Add include files that define the shape of the CAPTAIN inner and outer vessels.
This adds the infrastructure to take the digitized boundaries of a vessel
and turn it into the values needed to build a G4Polycone shape.  The
polycone segments are pushed into a vector with the Z, inner and outer
radius values.  The commit includes the c++ include files, defining the
polycone boundaries, the digitized files with the boundary, and the script
to translate the digitized files into the polycone boundary.