Much improved cryostat shapes for CAPTAIN and miniCAPTAIN
This include the digitized cryostat shapes for both CAPTAIN and
mini-CAPTAIN.  It also includes the updated PMT geometries for mini-CAPTAIN
and a guess at what will be used for CAPTAIN.  This is still missing bits
and pieces (e.g. the mini-CAPTAIN drift monitor), and needs to be validated
against as-built.