Rename the include files for the i/o classes to match ROOT expectations.
ROOT macros (.C) need to be able to find the i/o class include files, and
there are two conditions where the macros may be run.  In the first case,
they can be run as part of the edep-sim package, and the true include files
are available.  In the second, the include files are autogenerated as the
macro is run.  The source include files have been renamed so they match the
naming conventions for the autogenerated files.  This allows the ROOT
macros to run with either variety of file.