Tweak for large geometries.
The geometry can be built in two modes.  For small geometries, it is useful
to create a new TGeoNode for every object in the detector.  That makes it
more efficient to find the nodes and they can be referenced to refer to a
particular physical object.  For large geometries, there are to many nodes
so they all can't be created (and still exist in memory).  In that case,
the physical object will have to be referenced by the full path for the
node.  This is less compact and more inconvenient, but also more general.
The current transition is at 100,000 physical objects in the detector.