First addition of a very simple event display.
This isn't a real display, but it's intended to help debug the simulation
and simulated geometries.  It can also be taken as an example of how to use
the edep-sim output file.  This started life as a python script, but there
are still to many bugs in the pyROOT access to Eve for it to work.