Change VKinematicsGenerator::GeneratePrimaryVertex return value to an enum.
The return value has been changed so that it is a status enum instead of a
boolean.   The current valid values are kFail (equivalent to false),
kSuccess (equivalent to true), and kLast.  The kLast value will be used to
flag that the newly generated vertex should be the last vertex added to the
vertex by the generator.