Make the i/o fields private by default.
This begins deprecating the public fields in the i/o classes.  For now,
they remain public by default, but accessor methods are provided.  The
fields can be made private using the EDEPSIM_FORCE_PRIVATE_FIELDS macro.
In the future, the default will be for the fields to be private, and in
that case, they can be made public using the EDEPSIM_USE_PUBLIC_FIELDS
macro.