Make tweak to tickle a bug.
This makes a change to how a TPolyhedra solid is created so that it
tickles a bug in the G4 to ROOT geometry translation.  The
functionality should not have changed.  The current code will build,
but the produced ROOT geometry is incorrect.  The ROOT geometry will
be fixed in a subsequent commit.