Fix construction of TGeoPgon from G4Polyhedra
EDepSimRootGeometryManager has been changed to be less sensitive to
the order in which the Z sections are defined in the construction of a
G4Polyhedra solid.  Credit goes to Eldwan Brianne and Mike Kordowsky
for pointing out the bug, this is a modified version of their merge
requests.

This fixes how a ROOT TGeoPgon is constructed from a Geant4
G4Polyhedra.  The G4Polyhedra is can be defined by specifying
"sections" which specify the inner and outer radius of a polyhedral
solid at a particular z (in an "rz" coordinate system).  The
G4Polyhedra also allows the shape of the solid to be specified as a
path around the cross-section.

Internally to G4Polyhedra, the sections used to define the z,rmin,rmax
sections are turned into a path around a slice of the polyhedral solid
(in r,z).  The path is (apparently) tranversed in a counter clockwise
direction.  For slices, this means that the minimum radii, or the
maximum radii elements may be revereed in order. The order will depend
on whether the solid is specified from maxZ->minZ, or minZ->maxZ.