Initial commit of code to simulate with arbitrary magnetic fields.
New classes to handle arbitrary magnetic and electric fields and
interfacing to the rest of Geant4.

New class for interpolation routines for the field grids.

Edits to existing uniform field classes to work with the new method for
setting up a field.

Edits to UserDetectorConstruction to include arbitrary fields as an
option and handle constructing the proper EMField object for the field
manager.