Remove commented code, adjust variable names.
Renamed Bicubic to Cubic for the interpolation class.

Removed a number of commented out lines of code which were either better
implemented or were for testing/development.

Adjusted a number of variable names to better match the style of the
surrounding code (i.e. changing from snake_case to camelCase).