Add macro command to make volumes "insensitive"
Occasionally, a GDML file will mark large parts of the world as
sensitive, but that will make events very large during production
runs.  This adds a macro command to turn off sensitive volumes at
runtime.  The command is "/edep/hitExclude [LogicalVolume]", and it
needs to be set before the /edep/update command.