Add EDepSimUnits.h to document units used by EDepSim (same as G4)
The EDepSimUnits.h file is an extremely lightly edited version of the
GEANT4 version of the CLHEP SystemOfUnits.h file.  The only changes
are in the comments to clarify the history of the file (i.e. where it
was copied from), and to place the units into the EDepSim namespace
(instead of CLHEP).  This file falls under the GEANT4 license.